backend/requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
supabase==2.3.4
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.3.2
gunicorn==21.2.0
requests==2.31.0
